requirements.txt
audioop-lts==0.2.2
audioread==3.0.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.60.0
fsspec==2025.9.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.0
MarkupSafe==3.0.2
matplotlib==3.10.6
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numba==0.62.0
numpy==2.3.3
onnx==1.19.0
onnxruntime==1.22.1
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
protobuf==6.32.1
pycparser==2.23
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.13.3
threadpoolctl==3.6.0
torch==2.8.0
torchaudio==2.8.0
torchcodec==0.7.0
torchvision==0.23.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
